requirements.txt
langchain==0.2.0
langgraph==0.2.0
langchain-community==0.2.0
langchain-openai==0.1.7
requests==2.31.0
beautifulsoup4==4.12.0
python-dotenv==1.0.0
openai>=1.24.0
openmeteo-requests==1.1.0
requests-cache==1.1.1
retry-requests==2.0.0
matplotlib>=3.7.0
networkx>=3.1
wikipedia==1.4.0
